feat(layout): a11y skip-link + Suspense-defer sidebar 13-hook fan-out
Wave 3.5 + Wave 5 (session-G W1.A cache wrapper context).

PART A — a11y polish on existing skip-to-content link:
- Add `tabIndex={-1}` to the existing `<main id="main-content">` so
  the skip target reliably receives focus on activation. The link
  itself (focus:not-sr-only chrome) and `<html lang="en">` were
  already on main; this commit makes them work end-to-end for
  screen-reader / keyboard users.

PART B — defer the sidebar's 13-hook fan-out behind <Suspense>:
- `getPublicSidebarShell()` runs `getSidebarSourceCounts()` which
  refreshes 13 data sources (HN / Lobsters / Devto / Bluesky /
  Reddit / PH / funding / revenue / npm / skills / MCP / agent /
  HF). On a cold render the whole thing blocked the layout's RSC
  stream for ~4.2 s before `{children}` could paint.
- Extract `<SidebarStream />` (new server component) that owns the
  await + try/catch + `<Sidebar initialShell={shell}/>` render.
- Wrap it in `<Suspense fallback={<SidebarSkeleton />}>` inside
  `<AppShell>`. The page body now streams immediately; the
  sidebar resolves and patches in as a follow-up RSC flight.
- Failure mode unchanged: shell build throws -> initialShell=null
  -> client Sidebar falls back to its own /api/pipeline/sidebar-data
  fetch (already implemented in `useSidebarData`).

Files:
- src/app/layout.tsx (modified — Suspense wrap, removed inline await)
- src/components/layout/SidebarStream.tsx (new — async server component)

Verification:
- `npm run typecheck`            clean (0 errors)
- `npm run lint:guards`          clean (0 missing budgets, 0 orphaned)
